requirements.txt
# Core
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
loguru>=0.7.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0
pyarrow>=14.0.0

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
scrapy>=2.11.0
selenium>=4.15.0

# RAG and ML
openai>=1.3.0
anthropic>=0.39.0
langchain>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.12.0

# Azure
azure-identity>=1.15.0
azure-ai-ml>=1.14.0
azure-keyvault-secrets>=4.7.0
azure-storage-blob>=12.19.0

# Monitoring and logging
prometheus-client>=0.19.0
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
black>=23.11.0
ruff>=0.1.6
mypy>=1.7.0
pre-commit>=3.5.0
